% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWFisher.MultiPheno.R
\name{AWFisher.MultiPheno}
\alias{AWFisher.MultiPheno}
\title{Multivariate phenotype-gene association analysis}
\usage{
AWFisher.MultiPheno(
  expr,
  pheno,
  confounder,
  Pheno.type,
  method = c("AFp", "AFz"),
  num.perm = 100,
  num.bootstrap = 50,
  ncores = 1
)
}
\description{
Multivariate phenotype-gene association analysis
}
